Add check to make_miniboost if BOOST_ROOT is not set

+++ make_miniboost.sh
@@ -1,4 +1,9 @@
-
+#!/bin/sh
+if test -z "$BOOST_ROOT"; then
+  echo "Set BOOST_ROOT to your Boost directory!"
+  echo "If you use your system Boost, export BOOST_ROOT=/usr/include"
+  exit 1
+fi
 rm miniboost -rf && mkdir miniboost
 bcp --boost=$BOOST_ROOT program_options foreach thread miniboost
 cp -r $BOOST_ROOT/boost-build.jam  $BOOST_ROOT/boostcpp.jam $BOOST_ROOT/bootstrap.sh miniboost/
